1.13 (#193)
* refactor: start moving to obsidian 1.13

* Fix lint quote style mismatch (#192)

* Changes before error encountered

Agent-Logs-Url: https://github.com/DontBlameMe99/Tick-Tones/sessions/de36af53-291d-46d5-9bcf-b7e3be349a43

* Fix declarative settings typing

* fix: align lint quote style

---------


---------

+++ eslint.config.mjs
@@ -9,7 +9,7 @@ export default tseslint.config(
   ...tseslint.configs.recommended,
   {
     rules: {
-      quotes: ['error', 'double', { avoidEscape: true }],
+      quotes: ['error', 'single', { avoidEscape: true }],
       'jsx-quotes': ['error', 'prefer-double'],
       indent: [
         'error',
